refactor for name services
